grantflow-backend-py/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic[email]==2.6.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
requests==2.31.0
alembic==1.13.1
google-generativeai==0.8.5
